requirements.txt
livekit-agents[deepgram,openai,cartesia,silero,turn-detector]~=1.2
livekit-plugins-noise-cancellation~=0.2
python-dotenv
pinecone
deepgram-sdk
google-ai-generativelanguage
cartesia
silero
livekit-plugins-google
livekit-plugins-noise-cancellation
livekit-plugins-turn-detector




  